addressiong pr comments
